refactor: clean up control.cpp - move readLine to /etc/utils.cpp - seperate writePNG and readPNG from control.cpp - seperate i/o interaction from readInput
