+++ src/etc/utils.hpp
@@ -11,5 +11,6 @@ void flog(int x);
 void flogcolor(color c);
 int pythagoreanApprox(int a, int b);
 int ColorRGB(int R, int G, int B);
+char* readLine() ;
 
 #endif
